requirements.txt
requests
python-dotenv
tabulate
pyyaml
colorama
